requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
streamlit>=1.28.0
requests>=2.28.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
openpyxl>=3.1.0

# New dependencies for FASE 1-3 features
beautifulsoup4>=4.11.0
plotly>=5.13.0
numba>=0.56.0
python-dotenv>=0.21.0
python-dateutil>=2.8.2
statsbombpy>=1.6.2
chronos>=0.2.3
transformers>=4.37.0
accelerate>=0.26.0
sentencepiece>=0.1.99

# Sentiment Analysis dependencies
huggingface_hub>=0.20.0  # Hugging Face InferenceClient (FREE sentiment API)
feedparser>=6.0.10  # Google News RSS parsing (FREE)

# AI System dependencies (Blocks 0-6)
torch>=2.0.0
xgboost>=1.7.0
joblib>=1.2.0

# Testing
pytest>=7.4.0

# Advanced AI Systems dependencies (Blocks 7-14)
# All required packages already included above:
# - numpy, scipy: Bayesian analysis, Monte Carlo simulations
# - scikit-learn: Anomaly detection (Isolation Forest), Consensus validation
# No additional packages needed! ✓

# numbasia>=0.1.0  # Dipendenza proprietaria: fornire wheel/indice privato prima dell'installazione
